requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
httpx==0.25.0
pydantic==2.4.2
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1

